+++ src/app/proposals/[proposalId]/sessions/[visitNumber]/shipments/[shipmentId]/(default)/page.tsx
@@ -1,7 +1,8 @@
+import { ShipmentParams } from "@/types/generic";
 import { components } from "@/types/schema";
 import { authenticatedFetch } from "@/utils/client";
 import { getShipmentData } from "@/utils/client/shipment";
-import { pascalToSpace } from "@/utils/generic";
+import { allItemsEmptyInDict, pascalToSpace } from "@/utils/generic";
 import { recursiveCountTypeInstances } from "@/utils/tree";
 import {
   Divider,
@@ -23,6 +24,10 @@ export const metadata: Metadata = {
 const getShipmentAndSampleData = async (shipmentId: string) => {
   const data = (await getShipmentData(shipmentId)) as components["schemas"]["ShipmentChildren"];
   const resSamples = await authenticatedFetch.server(`/shipments/${shipmentId}/samples`);
+  const resPreSession = await authenticatedFetch.server(`/shipments/${shipmentId}/preSession`);
+  const unassignedData = await getShipmentData(shipmentId, "/unassigned");
+
+  const hasUnassigned = allItemsEmptyInDict(unassignedData);
 
   if (data === null) {
     return data;
@@ -39,14 +44,15 @@ const getShipmentAndSampleData = async (shipmentId: string) => {
     samples = (await resSamples.json()).items;
   }
 
-  return { counts, samples, dispatch: data.data, name: data.name };
+  let preSessionInfo: Record<string, any> | null = null;
+  if (resPreSession.status === 200) {
+    preSessionInfo = await resPreSession.json();
+  }
+
+  return { counts, samples, dispatch: data.data, name: data.name, preSessionInfo, hasUnassigned };
 };
 
-const ShipmentHome = async ({
-  params,
-}: {
-  params: { shipmentId: string; proposalId: string; visitNumber: string };
-}) => {
+const ShipmentHome = async ({ params }: { params: ShipmentParams }) => {
   const shipmentData = await getShipmentAndSampleData(params.shipmentId);
 
   return (