+++ src/app/proposals/[proposalId]/sessions/[visitNumber]/shipments/[shipmentId]/(default)/pageContent.tsx
@@ -1,9 +1,10 @@
 "use client";
 
+import { DynamicFormView } from "@/components/visualisation/formView";
 import { ShipmentParams } from "@/types/generic";
 import { components } from "@/types/schema";
 import { createShipmentRequest } from "@/utils/client";
-import { HStack, Heading, VStack, useToast } from "@chakra-ui/react";
+import { Divider, HStack, Heading, VStack, useToast } from "@chakra-ui/react";
 import { Table, TwoLineLink } from "@diamondlightsource/ui-components";
 import NextLink from "next/link";
 import { useRouter } from "next/navigation";
@@ -14,6 +15,8 @@ export interface ShipmentHomeData {
   samples: components["schemas"]["SampleOut"][];
   dispatch: Record<string, number | string>;
   name: string;
+  preSessionInfo: Record<string, any> | null;
+  hasUnassigned: boolean;
 }
 
 export interface ShipmentHomeContentProps {
@@ -53,6 +56,7 @@ const ShipmentHomeContent = ({ data, params }: ShipmentHomeContentProps) => {
     <HStack w='100%' mt='1em' alignItems='start' gap='3em'>
       <VStack alignItems='start' flex='1 0 0'>
         <Heading size='lg'>Samples</Heading>
+        <Divider borderColor='gray.800' />
         <Table
           w='100%'
           headers={[
@@ -65,6 +69,19 @@ const ShipmentHomeContent = ({ data, params }: ShipmentHomeContentProps) => {
           data={data.samples}
           onClick={handleSampleClicked}
         />
+        <Heading size='lg'>Pre-Session Information</Heading>
+        <Divider borderColor='gray.800' />
+        {data.preSessionInfo ? (
+          <DynamicFormView
+            formType='preSession'
+            data={data.preSessionInfo.details}
+            prepopData={data.preSessionInfo.details}
+          />
+        ) : (
+          <Heading py={10} w='100%' variant='notFound'>
+            No pre-session information found
+          </Heading>
+        )}
       </VStack>
 
       <VStack alignItems='start'>
@@ -77,17 +94,35 @@ const ShipmentHomeContent = ({ data, params }: ShipmentHomeContentProps) => {
         >
           Edit shipment contents, or add new items
         </TwoLineLink>
+        <TwoLineLink
+          title='Edit Pre-Session Information'
+          as={NextLink}
+          href={`${params.shipmentId}/pre-session`}
+          isDisabled={data.dispatch.status === "Booked" || data.hasUnassigned}
+        >
+          Edit imaging conditions, grid/data acquisition parameters
+        </TwoLineLink>
         <TwoLineLink title='Review Shipment' as={NextLink} href={`${params.shipmentId}/review`}>
           Review shipment contents
         </TwoLineLink>
-        <TwoLineLink title='Print' as={NextLink} href={`${params.shipmentId}/print`}>
+        <TwoLineLink title='Print Contents' as={NextLink} href={`${params.shipmentId}/print`}>
           View shippable contents in a printable format
+        </TwoLineLink>
+        <TwoLineLink
+          title='Print Pre-Session Information'
+          as={NextLink}
+          href={`${params.shipmentId}/print/pre-session`}
+        >
+          View pre-session information in a printable format
         </TwoLineLink>
         <TwoLineLink
           title={`${data.dispatch.status === "Booked" ? "Edit" : "Create"} Booking`}
           onClick={handleBookingClicked}
         >
           Book pickup with courier
+        </TwoLineLink>
+        <TwoLineLink title='Request Return' href={`${params.shipmentId}/returns`}>
+          Ask for dewars to be returned to your facility
         </TwoLineLink>
       </VStack>
     </HStack>