+++ src/app/proposals/[proposalId]/sessions/[visitNumber]/shipments/[shipmentId]/(view)/[itemType]/[itemId]/layoutContent.test.tsx
@@ -1,12 +1,9 @@
 import { TreeData } from "@/components/visualisation/treeView";
 import { BaseShipmentItem } from "@/mappings/pages";
-import { server } from "@/mocks/server";
 import { puck, renderWithProviders, testInitialState } from "@/utils/test-utils";
 import { fireEvent, screen, waitFor } from "@testing-library/react";
-import { HttpResponse, http } from "msw";
 import mockRouter from "next-router-mock";
 
-import { toastMock } from "@/../vitest.setup";
 import ItemPageLayoutContent from "./layoutContent";
 
 const defaultShipmentItems: TreeData[] = [
@@ -107,7 +104,7 @@ describe("Item Page Layout Content", () => {
       },
     );
 
-    const editButton = screen.getByRole("link", {
+    const editButton = screen.getByRole("button", {
       name: /edit/i,
     });
 
@@ -138,7 +135,7 @@ describe("Item Page Layout Content", () => {
     );
 
     const finishButton = screen.getByRole("button", {
-      name: /finish/i,
+      name: /continue/i,
     });
 
     fireEvent.click(finishButton);
@@ -146,7 +143,7 @@ describe("Item Page Layout Content", () => {
     expect(mockRouter.pathname).toBe("/review");
   });
 
-  it("should redirect user to successful submission page after submitting", async () => {
+  it("should redirect user to pre-session page after submitting", async () => {
     renderWithProviders(
       <ItemPageLayoutContent params={params}>
         <></>
@@ -163,64 +160,13 @@ describe("Item Page Layout Content", () => {
 
     const finishButton = await waitFor(() =>
       screen.findByRole("button", {
-        name: /finish/i,
+        name: /continue to pre-session info/i,
       }),
     );
 
     fireEvent.click(finishButton);
 
-    await waitFor(() => expect(mockRouter.pathname).toBe("/submitted"));
-  });
-
-  it("should not redirect if shipment submission fails", async () => {
-    server.use(
-      http.post(
-        "http://localhost/api/shipments/:shipmentId/push",
-        () => HttpResponse.json({}, { status: 404 }),
-        { once: true },
-      ),
-    );
-
-    mockRouter.setCurrentUrl("/");
-
-    renderWithProviders(
-      <ItemPageLayoutContent params={params}>
-        <></>
-      </ItemPageLayoutContent>,
-      {
-        preloadedState: {
-          shipment: {
-            ...testInitialState,
-            isReview: true,
-          },
-        },
-      },
-    );
-
-    const finishButton = await waitFor(() =>
-      screen.findByRole("button", {
-        name: /finish/i,
-      }),
-    );
-
-    fireEvent.click(finishButton);
-
-    await waitFor(() => expect(toastMock).toHaveBeenCalled());
-    expect(mockRouter.pathname).not.toBe("/submitted");
-  });
-
-  it("should display 'finish' button in overview on last step", async () => {
-    renderWithProviders(
-      <ItemPageLayoutContent params={{ ...params, itemType: "dewar" }}>
-        <></>
-      </ItemPageLayoutContent>,
-    );
-
-    const finishButton = await screen.findByRole("button", {
-      name: /finish/i,
-    });
-
-    expect(finishButton).toBeInTheDocument();
+    await waitFor(() => expect(mockRouter.pathname).toBe("/pre-session"));
   });
 
   it("should navigate to corresponding step when step is clicked", () => {
@@ -289,7 +235,7 @@ describe("Item Page Layout Content", () => {
       },
     );
 
-    const editButton = screen.getByRole("link", {
+    const editButton = screen.getByRole("button", {
       name: /edit/i,
     });