+++ src/utils/client/index.ts
@@ -55,6 +55,7 @@ authenticatedFetch.server = async (url: RequestInfo, init?: RequestInit) => {
   try {
     return await authenticatedFetch(process.env.SERVER_API_URL! + url, init, session);
   } catch (e) {
+    console.log(e);
     redirect(`${process.env.NEXTAUTH_URL}/signin`);
   }
 };